+++ clean_orphans_elements.py
@@ -1,9 +1,8 @@
 import argparse
 
-from functions.clean_orphan_contingency_lists import delete_orphan_contingency_lists
-from functions.clean_orphan_filters import delete_orphan_filters
-from functions.clean_orphan_networks import delete_orphan_network
-
+from functions.clean_orphan_elements.clean_orphan_contingency_lists import delete_orphan_contingency_lists
+from functions.clean_orphan_elements.clean_orphan_filters import delete_orphan_filters
+from functions.clean_orphan_elements.clean_orphan_networks import delete_orphan_network
 from script_mode import ScriptMode
 
 parser=argparse.ArgumentParser(description='Send requests to the gridsuite services to remove orphan elements',)
